Cleanup and refactor

+++ src/main/scala/fr/renoux/gaston/engine/Scoring.scala
@@ -5,13 +5,13 @@ import fr.renoux.gaston.model.preferences.Preference
 import fr.renoux.gaston.model.problem.Problem
 
 /**
-  * Created by gael on 07/05/17.
+  * Scoring a schedule for a certain problem.
   */
 object Scoring {
   val PreferencesScore = Map(Preference.Strong -> 10, Preference.Weak -> 1)
 
   /** Returns Left with the number (>0) of broken constraints, or Right with the score. */
-  def score(problem: Problem, schedule: Schedule): Either[Long, Long] = {
+  def score(problem: Problem, schedule: Schedule): Either[Int, Int] = {
     val constraintsBroken = problem.constraints.map(_.countBroken(schedule)).sum
     if (constraintsBroken > 0) Left(constraintsBroken)
     else {
